requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
asyncer==0.0.8
attrs==25.3.0
audioop-lts==0.2.2
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
bidict==0.23.1
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chainlit==2.7.2
charset-normalizer==3.4.3
chevron==0.14.0
chromadb==1.0.20
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.7
cuid==0.4
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
durationpy==0.10
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.9.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
gpt4all==2.8.2
greenlet==3.2.4
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.29
langchain-core==0.3.75
langchain-ollama==0.3.7
langchain-text-splitters==0.3.11
langchainhub==0.1.21
langsmith==0.4.27
Lazify==0.4.0
literalai==0.1.201
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.13.1
mdurl==0.1.2
mmh3==5.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.2
oauthlib==3.3.1
ollama==0.5.3
onnxruntime==1.22.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-alephalpha==0.47.0
opentelemetry-instrumentation-anthropic==0.47.0
opentelemetry-instrumentation-bedrock==0.47.0
opentelemetry-instrumentation-chromadb==0.47.0
opentelemetry-instrumentation-cohere==0.47.0
opentelemetry-instrumentation-crewai==0.47.0
opentelemetry-instrumentation-google-generativeai==0.47.0
opentelemetry-instrumentation-groq==0.47.0
opentelemetry-instrumentation-haystack==0.47.0
opentelemetry-instrumentation-lancedb==0.47.0
opentelemetry-instrumentation-langchain==0.47.0
opentelemetry-instrumentation-llamaindex==0.47.0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-marqo==0.47.0
opentelemetry-instrumentation-mcp==0.47.0
opentelemetry-instrumentation-milvus==0.47.0
opentelemetry-instrumentation-mistralai==0.47.0
opentelemetry-instrumentation-ollama==0.47.0
opentelemetry-instrumentation-openai==0.47.0
opentelemetry-instrumentation-openai-agents==0.47.0
opentelemetry-instrumentation-pinecone==0.47.0
opentelemetry-instrumentation-qdrant==0.47.0
opentelemetry-instrumentation-redis==0.57b0
opentelemetry-instrumentation-replicate==0.47.0
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-sagemaker==0.47.0
opentelemetry-instrumentation-sqlalchemy==0.57b0
opentelemetry-instrumentation-threading==0.57b0
opentelemetry-instrumentation-together==0.47.0
opentelemetry-instrumentation-transformers==0.47.0
opentelemetry-instrumentation-urllib3==0.57b0
opentelemetry-instrumentation-vertexai==0.47.0
opentelemetry-instrumentation-watsonx==0.47.0
opentelemetry-instrumentation-weaviate==0.47.0
opentelemetry-instrumentation-writer==0.47.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.57b0
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
posthog==3.25.0
propcache==0.3.2
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.0.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.3
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
tokenizers==0.22.0
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.47.0
typer==0.17.4
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
